feat: [CI-3648]: refactor for sonar
