+++ src/modules/70-pipeline/utils/constants.ts
@@ -7,7 +7,6 @@
 import type { IconName, SelectOption } from '@harness/uicore'
 import type { ConnectorInfoDTO } from 'services/cd-ng'
 
-import type { StepViewType } from '@pipeline/components/AbstractSteps/Step'
 import type { UseStringsReturn } from 'framework/strings'
 
 export const EXPRESSION_STRING = '<+expression>' // TODO: this needs to be exported from uicore for best use.
@@ -40,9 +39,6 @@ export const stageTypeToIconMap: Record<string, IconName> = {
   Approval: 'approval-stage-icon',
   FeatureFlag: 'cf-main'
 }
-
-export const getConnectorRefWidth = (viewType: StepViewType): number =>
-  Object.entries(ConnectorRefWidth).find(key => key[0] === viewType)?.[1] || ConnectorRefWidth.DefaultView
 
 export const getPrCloneStrategyOptions = (getString: UseStringsReturn['getString']): SelectOption[] => [
   { label: getString('pipeline.rightBar.mergeCommit'), value: 'MergeCommit' }, // should keep as index 0 for default value
